Extract SharedDict to separate file
Move SharedDict implementation from py_nif.c and py_callback.c
to dedicated py_shared_dict.c for better code organization.
