+++ c_src/py_callback.c
@@ -3281,365 +3281,6 @@ static PyObject *erlang_whereis_impl(PyObject *self, PyObject *args) {
     return result;
 }
 
-/* ============================================================================
- * SharedDict Python Methods
- *
- * These functions implement the Python-side API for SharedDict, allowing
- * Python code to access process-scoped shared dictionaries.
- * ============================================================================ */
-
-/**
- * Extract py_shared_dict_t* from a PyCapsule.
- * Returns NULL and sets exception on error.
- */
-static py_shared_dict_t *get_shared_dict_from_capsule(PyObject *capsule) {
-    if (!PyCapsule_IsValid(capsule, "py_shared_dict")) {
-        PyErr_SetString(PyExc_TypeError, "Invalid SharedDict handle");
-        return NULL;
-    }
-    py_shared_dict_t *sd = (py_shared_dict_t *)PyCapsule_GetPointer(capsule, "py_shared_dict");
-    if (sd == NULL) {
-        PyErr_SetString(PyExc_ValueError, "SharedDict handle is NULL");
-        return NULL;
-    }
-    if (atomic_load(&sd->destroyed)) {
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-    return sd;
-}
-
-/**
- * Python implementation of SharedDict.get(key, default=None)
- * Usage: erlang._shared_dict_get(handle, key)
- */
-static PyObject *py_shared_dict_get_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *handle;
-    const char *key;
-    Py_ssize_t key_len;
-
-    if (!PyArg_ParseTuple(args, "Os#", &handle, &key, &key_len)) {
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = get_shared_dict_from_capsule(handle);
-    if (sd == NULL) return NULL;
-
-    pthread_mutex_lock(&sd->mutex);
-
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-
-    /* Create Python key from bytes */
-    PyObject *py_key = PyBytes_FromStringAndSize(key, key_len);
-    if (py_key == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    /* Look up pickled value */
-    PyObject *pickled = PyDict_GetItem(sd->dict, py_key);
-    Py_DECREF(py_key);
-
-    if (pickled == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        Py_RETURN_NONE;
-    }
-
-    /* Unpickle the value */
-    PyObject *pickle_mod = PyImport_ImportModule("pickle");
-    if (pickle_mod == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    PyObject *loads = PyObject_GetAttrString(pickle_mod, "loads");
-    Py_DECREF(pickle_mod);
-    if (loads == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    PyObject *value = PyObject_CallFunctionObjArgs(loads, pickled, NULL);
-    Py_DECREF(loads);
-
-    pthread_mutex_unlock(&sd->mutex);
-    return value;
-}
-
-/**
- * Python implementation of SharedDict.set(key, value)
- * Usage: erlang._shared_dict_set(handle, key, value)
- */
-static PyObject *py_shared_dict_set_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *handle;
-    const char *key;
-    Py_ssize_t key_len;
-    PyObject *value;
-
-    if (!PyArg_ParseTuple(args, "Os#O", &handle, &key, &key_len, &value)) {
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = get_shared_dict_from_capsule(handle);
-    if (sd == NULL) return NULL;
-
-    pthread_mutex_lock(&sd->mutex);
-
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-
-    /* Pickle the value */
-    PyObject *pickle_mod = PyImport_ImportModule("pickle");
-    if (pickle_mod == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    PyObject *dumps = PyObject_GetAttrString(pickle_mod, "dumps");
-    Py_DECREF(pickle_mod);
-    if (dumps == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    PyObject *pickled = PyObject_CallFunctionObjArgs(dumps, value, NULL);
-    Py_DECREF(dumps);
-    if (pickled == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    /* Create Python key */
-    PyObject *py_key = PyBytes_FromStringAndSize(key, key_len);
-    if (py_key == NULL) {
-        Py_DECREF(pickled);
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    /* Store in dict */
-    int result = PyDict_SetItem(sd->dict, py_key, pickled);
-    Py_DECREF(py_key);
-    Py_DECREF(pickled);
-
-    pthread_mutex_unlock(&sd->mutex);
-
-    if (result < 0) {
-        return NULL;
-    }
-    Py_RETURN_NONE;
-}
-
-/**
- * Python implementation of SharedDict.delete(key)
- * Usage: erlang._shared_dict_del(handle, key)
- * Returns True if key existed, False otherwise
- */
-static PyObject *py_shared_dict_del_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *handle;
-    const char *key;
-    Py_ssize_t key_len;
-
-    if (!PyArg_ParseTuple(args, "Os#", &handle, &key, &key_len)) {
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = get_shared_dict_from_capsule(handle);
-    if (sd == NULL) return NULL;
-
-    pthread_mutex_lock(&sd->mutex);
-
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-
-    /* Create Python key */
-    PyObject *py_key = PyBytes_FromStringAndSize(key, key_len);
-    if (py_key == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    /* Check if key exists */
-    int exists = PyDict_Contains(sd->dict, py_key);
-    if (exists > 0) {
-        PyDict_DelItem(sd->dict, py_key);
-        PyErr_Clear();
-    }
-    Py_DECREF(py_key);
-
-    pthread_mutex_unlock(&sd->mutex);
-
-    if (exists > 0) {
-        Py_RETURN_TRUE;
-    }
-    Py_RETURN_FALSE;
-}
-
-/**
- * Python implementation of SharedDict.contains(key)
- * Usage: erlang._shared_dict_contains(handle, key)
- */
-static PyObject *py_shared_dict_contains_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *handle;
-    const char *key;
-    Py_ssize_t key_len;
-
-    if (!PyArg_ParseTuple(args, "Os#", &handle, &key, &key_len)) {
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = get_shared_dict_from_capsule(handle);
-    if (sd == NULL) return NULL;
-
-    pthread_mutex_lock(&sd->mutex);
-
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-
-    /* Create Python key */
-    PyObject *py_key = PyBytes_FromStringAndSize(key, key_len);
-    if (py_key == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    int exists = PyDict_Contains(sd->dict, py_key);
-    Py_DECREF(py_key);
-
-    pthread_mutex_unlock(&sd->mutex);
-
-    if (exists > 0) {
-        Py_RETURN_TRUE;
-    } else if (exists == 0) {
-        Py_RETURN_FALSE;
-    } else {
-        return NULL;  /* Error occurred */
-    }
-}
-
-/**
- * Python implementation of SharedDict.destroy()
- * Usage: erlang._shared_dict_destroy(handle)
- *
- * Explicitly destroys the SharedDict, invalidating all references.
- * This is idempotent - calling on already-destroyed dict returns None.
- */
-static PyObject *py_shared_dict_destroy_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *capsule;
-
-    if (!PyArg_ParseTuple(args, "O", &capsule)) {
-        return NULL;
-    }
-
-    if (!PyCapsule_IsValid(capsule, "py_shared_dict")) {
-        PyErr_SetString(PyExc_TypeError, "Invalid SharedDict handle");
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = (py_shared_dict_t *)PyCapsule_GetPointer(capsule, "py_shared_dict");
-    if (sd == NULL) {
-        PyErr_SetString(PyExc_ValueError, "SharedDict handle is NULL");
-        return NULL;
-    }
-
-    /* If already destroyed, return None (idempotent) */
-    if (atomic_load(&sd->destroyed)) {
-        Py_RETURN_NONE;
-    }
-
-    /* Mark as destroyed */
-    atomic_store(&sd->destroyed, true);
-
-    /* Clear the Python dict */
-    pthread_mutex_lock(&sd->mutex);
-    Py_CLEAR(sd->dict);
-    pthread_mutex_unlock(&sd->mutex);
-
-    Py_RETURN_NONE;
-}
-
-/**
- * Python implementation of SharedDict.keys()
- * Usage: erlang._shared_dict_keys(handle)
- */
-static PyObject *py_shared_dict_keys_impl(PyObject *self, PyObject *args) {
-    (void)self;
-    PyObject *handle;
-
-    if (!PyArg_ParseTuple(args, "O", &handle)) {
-        return NULL;
-    }
-
-    py_shared_dict_t *sd = get_shared_dict_from_capsule(handle);
-    if (sd == NULL) return NULL;
-
-    pthread_mutex_lock(&sd->mutex);
-
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        PyErr_SetString(PyExc_RuntimeError, "SharedDict has been destroyed");
-        return NULL;
-    }
-
-    /* Get keys and decode them to strings */
-    PyObject *keys = PyDict_Keys(sd->dict);
-    if (keys == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    Py_ssize_t len = PyList_Size(keys);
-    PyObject *result = PyList_New(len);
-    if (result == NULL) {
-        Py_DECREF(keys);
-        pthread_mutex_unlock(&sd->mutex);
-        return NULL;
-    }
-
-    for (Py_ssize_t i = 0; i < len; i++) {
-        PyObject *py_key = PyList_GetItem(keys, i);
-        if (PyBytes_Check(py_key)) {
-            /* Decode bytes to string */
-            PyObject *str_key = PyUnicode_DecodeUTF8(
-                PyBytes_AS_STRING(py_key),
-                PyBytes_GET_SIZE(py_key),
-                "replace");
-            if (str_key == NULL) {
-                Py_DECREF(result);
-                Py_DECREF(keys);
-                pthread_mutex_unlock(&sd->mutex);
-                return NULL;
-            }
-            PyList_SET_ITEM(result, i, str_key);
-        } else {
-            Py_INCREF(py_key);
-            PyList_SET_ITEM(result, i, py_key);
-        }
-    }
-
-    Py_DECREF(keys);
-    pthread_mutex_unlock(&sd->mutex);
-    return result;
-}
-
 /* Python method definitions for erlang module */
 static PyMethodDef ErlangModuleMethods[] = {
     {"call", erlang_call_impl, METH_VARARGS,
