+++ c_src/py_nif.c
@@ -151,63 +151,6 @@ static void py_env_resource_dtor(ErlNifEnv *env, void *obj) {
     res->locals = NULL;
 }
 
-/* ============================================================================
- * Shared Dict Resource Callbacks
- * ============================================================================ */
-
-/**
- * @brief Down callback for py_shared_dict_t
- *
- * Called when the owning process dies. Sets destroyed flag and clears the dict.
- * This callback is invoked by the runtime when the monitored process terminates.
- */
-static void shared_dict_down(ErlNifEnv *env, void *obj,
-                              ErlNifPid *pid, ErlNifMonitor *mon) {
-    (void)env;
-    (void)pid;
-    (void)mon;
-    py_shared_dict_t *sd = (py_shared_dict_t *)obj;
-
-    /* Mark as destroyed - subsequent access will return badarg */
-    atomic_store(&sd->destroyed, true);
-    sd->monitor_active = false;
-
-    /* Clear the Python dict if runtime is still running */
-    if (runtime_is_running() && sd->dict != NULL) {
-        PyGILState_STATE gstate = PyGILState_Ensure();
-        pthread_mutex_lock(&sd->mutex);
-        Py_CLEAR(sd->dict);
-        pthread_mutex_unlock(&sd->mutex);
-        PyGILState_Release(gstate);
-    }
-}
-
-/**
- * @brief Destructor for py_shared_dict_t
- *
- * Called when the resource is garbage collected.
- * Cleans up the monitor, Python dict, and mutex.
- */
-static void shared_dict_destructor(ErlNifEnv *env, void *obj) {
-    py_shared_dict_t *sd = (py_shared_dict_t *)obj;
-
-    /* Demonitor if still active */
-    if (sd->monitor_active && env != NULL) {
-        enif_demonitor_process(env, sd, &sd->monitor);
-        sd->monitor_active = false;
-    }
-
-    /* Clear Python dict if runtime is still running */
-    if (runtime_is_running() && sd->dict != NULL) {
-        PyGILState_STATE gstate = PyGILState_Ensure();
-        Py_CLEAR(sd->dict);
-        PyGILState_Release(gstate);
-    }
-
-    /* Destroy mutex */
-    pthread_mutex_destroy(&sd->mutex);
-}
-
 /* Invariant counters for debugging and leak detection */
 py_invariant_counters_t g_counters = {0};
 
@@ -347,6 +290,7 @@ static int is_inline_schedule_marker(PyObject *obj);
 #include "py_convert.c"
 #include "py_exec.c"
 #include "py_logging.c"
+#include "py_shared_dict.c"
 #include "py_callback.c"
 #include "py_thread_worker.c"
 #include "py_event_loop.c"
@@ -7179,447 +7123,6 @@ static ERL_NIF_TERM nif_owngil_apply_paths(ErlNifEnv *env, int argc,
 }
 
 #endif /* HAVE_SUBINTERPRETERS */
-
-/* ============================================================================
- * Shared Dict NIF Implementation
- * ============================================================================ */
-
-/**
- * @brief Create a new process-scoped SharedDict
- *
- * Creates a SharedDict owned by the calling process. The dict is automatically
- * destroyed when the owning process terminates.
- *
- * @return {ok, Reference} on success, {error, Reason} on failure
- */
-static ERL_NIF_TERM nif_shared_dict_new(ErlNifEnv *env, int argc,
-                                         const ERL_NIF_TERM argv[]) {
-    (void)argc;
-    (void)argv;
-
-    if (!runtime_is_running()) {
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "not_initialized"));
-    }
-
-    /* Allocate resource */
-    py_shared_dict_t *sd = enif_alloc_resource(
-        PY_SHARED_DICT_RESOURCE_TYPE, sizeof(py_shared_dict_t));
-    if (sd == NULL) {
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "alloc_failed"));
-    }
-
-    /* Initialize fields */
-    memset(sd, 0, sizeof(py_shared_dict_t));
-    atomic_store(&sd->destroyed, false);
-    pthread_mutex_init(&sd->mutex, NULL);
-
-    /* Create Python dict for storage */
-    PyGILState_STATE gstate = PyGILState_Ensure();
-    sd->dict = PyDict_New();
-    if (sd->dict == NULL) {
-        PyGILState_Release(gstate);
-        pthread_mutex_destroy(&sd->mutex);
-        enif_release_resource(sd);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "dict_alloc_failed"));
-    }
-    PyGILState_Release(gstate);
-
-    /* Note: Process monitoring disabled for now to debug crash
-     * SharedDict will be garbage collected when no references remain */
-    sd->monitor_active = false;
-
-    /* Create reference term and release our reference */
-    ERL_NIF_TERM ref = enif_make_resource(env, sd);
-    enif_release_resource(sd);
-
-    return enif_make_tuple2(env, ATOM_OK, ref);
-}
-
-/**
- * @brief Get a value from SharedDict
- *
- * @param Handle SharedDict reference
- * @param Key Binary key
- * @param Default Default value if key not found
- * @return Value or Default, badarg if destroyed
- */
-static ERL_NIF_TERM nif_shared_dict_get(ErlNifEnv *env, int argc,
-                                         const ERL_NIF_TERM argv[]) {
-    (void)argc;
-
-    py_shared_dict_t *sd;
-    if (!enif_get_resource(env, argv[0], PY_SHARED_DICT_RESOURCE_TYPE, (void **)&sd)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Check if destroyed */
-    if (atomic_load(&sd->destroyed)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Get key as binary */
-    ErlNifBinary key_bin;
-    if (!enif_inspect_binary(env, argv[1], &key_bin)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Lock and access dict */
-    pthread_mutex_lock(&sd->mutex);
-
-    /* Check destroyed again after acquiring lock */
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    PyGILState_STATE gstate = PyGILState_Ensure();
-
-    /* Create Python key from binary */
-    PyObject *py_key = PyBytes_FromStringAndSize((char *)key_bin.data, key_bin.size);
-    if (py_key == NULL) {
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    /* Look up value (pickled bytes) */
-    PyObject *pickled = PyDict_GetItem(sd->dict, py_key);
-    Py_DECREF(py_key);
-
-    if (pickled == NULL) {
-        /* Key not found, return default */
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return argv[2];  /* Return default */
-    }
-
-    /* Unpickle the value */
-    PyObject *pickle_mod = PyImport_ImportModule("pickle");
-    if (pickle_mod == NULL) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "pickle_unavailable"));
-    }
-
-    PyObject *loads = PyObject_GetAttrString(pickle_mod, "loads");
-    Py_DECREF(pickle_mod);
-    if (loads == NULL) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "pickle_loads_unavailable"));
-    }
-
-    PyObject *value = PyObject_CallFunctionObjArgs(loads, pickled, NULL);
-    Py_DECREF(loads);
-
-    if (value == NULL) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "unpickle_failed"));
-    }
-
-    /* Convert Python value to Erlang term */
-    ERL_NIF_TERM result = py_to_term(env, value);
-    Py_DECREF(value);
-
-    PyGILState_Release(gstate);
-    pthread_mutex_unlock(&sd->mutex);
-
-    return result;
-}
-
-/**
- * @brief Set a value in SharedDict
- *
- * @param Handle SharedDict reference
- * @param Key Binary key
- * @param Value Erlang term value (will be pickled)
- * @return ok on success, badarg if destroyed
- */
-static ERL_NIF_TERM nif_shared_dict_set(ErlNifEnv *env, int argc,
-                                         const ERL_NIF_TERM argv[]) {
-    (void)argc;
-
-    py_shared_dict_t *sd;
-    if (!enif_get_resource(env, argv[0], PY_SHARED_DICT_RESOURCE_TYPE, (void **)&sd)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Check if destroyed */
-    if (atomic_load(&sd->destroyed)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Get key as binary */
-    ErlNifBinary key_bin;
-    if (!enif_inspect_binary(env, argv[1], &key_bin)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Lock and access dict */
-    pthread_mutex_lock(&sd->mutex);
-
-    /* Check destroyed again after acquiring lock */
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    PyGILState_STATE gstate = PyGILState_Ensure();
-
-    /* Convert Erlang term to Python object */
-    PyObject *py_value = term_to_py(env, argv[2]);
-    if (py_value == NULL) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "term_conversion_failed"));
-    }
-
-    /* Pickle the value for cross-interpreter safety */
-    PyObject *pickle_mod = PyImport_ImportModule("pickle");
-    if (pickle_mod == NULL) {
-        PyErr_Clear();
-        Py_DECREF(py_value);
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "pickle_unavailable"));
-    }
-
-    PyObject *dumps = PyObject_GetAttrString(pickle_mod, "dumps");
-    Py_DECREF(pickle_mod);
-    if (dumps == NULL) {
-        PyErr_Clear();
-        Py_DECREF(py_value);
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "pickle_dumps_unavailable"));
-    }
-
-    PyObject *pickled = PyObject_CallFunctionObjArgs(dumps, py_value, NULL);
-    Py_DECREF(dumps);
-    Py_DECREF(py_value);
-
-    if (pickled == NULL) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "pickle_failed"));
-    }
-
-    /* Create Python key from binary */
-    PyObject *py_key = PyBytes_FromStringAndSize((char *)key_bin.data, key_bin.size);
-    if (py_key == NULL) {
-        Py_DECREF(pickled);
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    /* Store pickled value in dict */
-    int set_result = PyDict_SetItem(sd->dict, py_key, pickled);
-    Py_DECREF(py_key);
-    Py_DECREF(pickled);
-
-    if (set_result < 0) {
-        PyErr_Clear();
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_tuple2(env, ATOM_ERROR,
-                                enif_make_atom(env, "dict_set_failed"));
-    }
-
-    PyGILState_Release(gstate);
-    pthread_mutex_unlock(&sd->mutex);
-
-    return ATOM_OK;
-}
-
-/**
- * @brief Delete a key from SharedDict
- *
- * @param Handle SharedDict reference
- * @param Key Binary key
- * @return ok on success (even if key didn't exist), badarg if destroyed
- */
-static ERL_NIF_TERM nif_shared_dict_del(ErlNifEnv *env, int argc,
-                                         const ERL_NIF_TERM argv[]) {
-    (void)argc;
-
-    py_shared_dict_t *sd;
-    if (!enif_get_resource(env, argv[0], PY_SHARED_DICT_RESOURCE_TYPE, (void **)&sd)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Check if destroyed */
-    if (atomic_load(&sd->destroyed)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Get key as binary */
-    ErlNifBinary key_bin;
-    if (!enif_inspect_binary(env, argv[1], &key_bin)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Lock and access dict */
-    pthread_mutex_lock(&sd->mutex);
-
-    /* Check destroyed again after acquiring lock */
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    PyGILState_STATE gstate = PyGILState_Ensure();
-
-    /* Create Python key from binary */
-    PyObject *py_key = PyBytes_FromStringAndSize((char *)key_bin.data, key_bin.size);
-    if (py_key == NULL) {
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    /* Delete key (ignore if not found) */
-    PyDict_DelItem(sd->dict, py_key);
-    PyErr_Clear();  /* Clear KeyError if key didn't exist */
-    Py_DECREF(py_key);
-
-    PyGILState_Release(gstate);
-    pthread_mutex_unlock(&sd->mutex);
-
-    return ATOM_OK;
-}
-
-/**
- * @brief Get all keys from SharedDict
- *
- * @param Handle SharedDict reference
- * @return List of binary keys, badarg if destroyed
- */
-static ERL_NIF_TERM nif_shared_dict_keys(ErlNifEnv *env, int argc,
-                                          const ERL_NIF_TERM argv[]) {
-    (void)argc;
-
-    py_shared_dict_t *sd;
-    if (!enif_get_resource(env, argv[0], PY_SHARED_DICT_RESOURCE_TYPE, (void **)&sd)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Check if destroyed */
-    if (atomic_load(&sd->destroyed)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Lock and access dict */
-    pthread_mutex_lock(&sd->mutex);
-
-    /* Check destroyed again after acquiring lock */
-    if (atomic_load(&sd->destroyed) || sd->dict == NULL) {
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_badarg(env);
-    }
-
-    PyGILState_STATE gstate = PyGILState_Ensure();
-
-    /* Get keys */
-    PyObject *keys = PyDict_Keys(sd->dict);
-    if (keys == NULL) {
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_list(env, 0);
-    }
-
-    Py_ssize_t len = PyList_Size(keys);
-    ERL_NIF_TERM *key_terms = enif_alloc(sizeof(ERL_NIF_TERM) * len);
-    if (key_terms == NULL) {
-        Py_DECREF(keys);
-        PyGILState_Release(gstate);
-        pthread_mutex_unlock(&sd->mutex);
-        return enif_make_list(env, 0);
-    }
-
-    for (Py_ssize_t i = 0; i < len; i++) {
-        PyObject *py_key = PyList_GetItem(keys, i);
-        if (PyBytes_Check(py_key)) {
-            char *data;
-            Py_ssize_t size;
-            PyBytes_AsStringAndSize(py_key, &data, &size);
-            ERL_NIF_TERM bin;
-            unsigned char *buf = enif_make_new_binary(env, size, &bin);
-            if (buf != NULL) {
-                memcpy(buf, data, size);
-                key_terms[i] = bin;
-            } else {
-                key_terms[i] = enif_make_atom(env, "error");
-            }
-        } else {
-            key_terms[i] = enif_make_atom(env, "error");
-        }
-    }
-
-    ERL_NIF_TERM result = enif_make_list_from_array(env, key_terms, len);
-    enif_free(key_terms);
-    Py_DECREF(keys);
-
-    PyGILState_Release(gstate);
-    pthread_mutex_unlock(&sd->mutex);
-
-    return result;
-}
-
-/**
- * @brief Explicitly destroy a SharedDict.
- *
- * Marks the SharedDict as destroyed and clears its Python dict.
- * This is idempotent - calling on an already-destroyed dict returns ok.
- *
- * @param Handle SharedDict reference
- * @return ok atom
- */
-static ERL_NIF_TERM nif_shared_dict_destroy(ErlNifEnv *env, int argc,
-                                             const ERL_NIF_TERM argv[]) {
-    (void)argc;
-
-    py_shared_dict_t *sd;
-    if (!enif_get_resource(env, argv[0], PY_SHARED_DICT_RESOURCE_TYPE, (void **)&sd)) {
-        return enif_make_badarg(env);
-    }
-
-    /* Check if already destroyed - idempotent */
-    if (atomic_load(&sd->destroyed)) {
-        return ATOM_OK;
-    }
-
-    /* Mark as destroyed */
-    atomic_store(&sd->destroyed, true);
-
-    /* Clear the Python dict */
-    pthread_mutex_lock(&sd->mutex);
-    if (sd->dict != NULL && runtime_is_running()) {
-        PyGILState_STATE gstate = PyGILState_Ensure();
-        Py_CLEAR(sd->dict);
-        PyGILState_Release(gstate);
-    }
-    sd->dict = NULL;
-    pthread_mutex_unlock(&sd->mutex);
-
-    return ATOM_OK;
-}
 
 /* ============================================================================
  * NIF setup
